+++ lib/runner/alex_command_runner.dart
@@ -4,6 +4,7 @@ import 'package:alex/commands/hello_world.dart';
 import 'package:alex/commands/l10n/l10n_command.dart';
 import 'package:alex/commands/pubspec/pubspec_command.dart';
 import 'package:alex/commands/release/release_command.dart';
+import 'package:alex/commands/settings/settings_command.dart';
 import 'package:alex/runner/alex_command.dart';
 import 'package:args/command_runner.dart';
 
@@ -18,6 +19,7 @@ class AlexCommandRunner extends CommandRunner<int> {
       CodeCommand(),
       PubspecCommand(),
       FeatureCommand(),
+      SettingsCommand(),
     ].forEach(addCommand);
   }
 }